refactor entity IDs from long to Integer and update repository interfaces accordingly; add MySQL configuration and initial data script

+++ src/main/java/com/br/mypoint/mypoint_skate/domain/repositories/PointRepository.java
@@ -1,11 +1,12 @@
 package com.br.mypoint.mypoint_skate.domain.repositories;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.br.mypoint.mypoint_skate.domain.models.Point;
 
 @Repository
-public interface PointRepository extends JpaRepository<Point, Long> {
+public interface PointRepository extends JpaRepository<Point, Integer> {
     
 }
